Extract `Push` and `Dup` into their own modules
This pulls those into their own modules, and gets rid of all the other logic in `common/mod.rs` since that has or will all be moved into typed instructions like `int/mod.rs`.
